% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strwh.R
\name{strwh}
\alias{strwh}
\title{Title}
\usage{
strwh(
  p,
  s = "Ahkhds",
  units = "mm",
  family = "serif",
  fontsize = 100,
  fontface = 1,
  familynames = NULL,
  work_path = NULL,
  tmp
)
}
\arguments{
\item{p}{str, "ggplot2" or "grib"}

\item{s}{str, a measured text}

\item{units}{str, "mm"}

\item{family}{str, "serif"}

\item{fontsize}{int, fontsize in ggplot2}

\item{fontface}{int, fontface in 1,2,3,4}

\item{familynames}{str, data/familynames.csv}

\item{tmp}{class}
}
\value{
a width of font, float,
}
\description{
Title
}
