% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strwh.R
\name{strwh_init}
\alias{strwh_init}
\title{Load python in use_python}
\usage{
strwh_init(
  work_path = NULL,
  python_path = NULL,
  fonts_path = NULL,
  FontFileType = "ttf"
)
}
\arguments{
\item{python_path}{str, a path of python.}

\item{fonts_path}{str, path of fonts}

\item{FontFileType}{str, ttf or afm}
}
\value{
a class
}
\description{
Load python in use_python
}
