% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strwh.R
\name{strwh_init}
\alias{strwh_init}
\title{Load python in use_python}
\usage{
strwh_init(python_path = "python")
}
\arguments{
\item{python_path}{str, a path of python.}
}
\value{
nothing
}
\description{
Load python in use_python
}
